requirements.txt
openai
whisper
sounddevice
numpy
elevenlabs 
PyQt6
pvporcupine 
python-dotenv
google-generativeai
rapidfuzz
pyttsx3
SpeechRecognition
pyaudio
requests
beautifulsoup4
pillow
pytesseract
pycaw
yfinance
pyautogui
screen-brightness-control
psutil
winshell
pypiwin32